Move models to another directory
