--- product-api/src/main/java/br/com/productapi/models/Category.java
@@ -1,18 +1,0 @@
-package br.com.productapi.models;
-
-import lombok.Data;
-
-import javax.persistence.*;
-
-@Data
-@Entity
-@Table(name = "category")
-public class Category{
-
-    @Id
-    @GeneratedValue(strategy= GenerationType.SEQUENCE)
-    private Integer id;
-
-    @Column(name = "description", nullable = false)
-    private String description;
-}
